pyproject.toml
[tool.poetry]
name = "exchange_cli_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
autoflake = "^1.3.1"
isort = "^4.3.21"
black = "^19.10b0"
unify = "^0.5"
flake8 = "^3.7.9"
pylint = "^2.4.4"
flask = "^1.1.1"
mypy = "^0.761"
sqlalchemy = "^1.3.15"
werkzeug = "0.16.1"
flask-restplus = "^0.13.0"
simplejson = "^3.17.0"
mock = "^4.0.2"
pytest-mock = "^2.0.0"
requests-mock = "^1.7.0"
pydantic = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"